\name{open-meta-analysis}
\alias{open-meta-analysis}
\title{open-meta-analysis}
\description{
This collection of functions implements  
\url{http://cran.r-project.org/web/packages/rmeta} at 
\url{https://public.opencpu.org}.

This application uses the rmeta package for R:

  Lumley, T. (2012). rmeta: Meta-analysis. Available at \url{http://cran.r-project.org/web/packages/rmeta/}.
	
}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\seealso{
\code{\link{open-meta-analysis}}.
}